feat: upgrade app packages
