+++ .gitignore
@@ -146,6 +146,7 @@ DerivedData
 
 # Fastlane
 fastlane/report.xml
+fastlane/Preview.html
 
 # symbols
 *.app.dSYM.zip
