+++ apollos.json
@@ -1,4 +1,4 @@
 {
-    "version": "2.36.3",
+    "version": "2.42.0",
     "environment": "client"
 }
