+++ fastlane/Fastfile
@@ -34,6 +34,8 @@ platform :ios do
   desc "Generate new provisioning profiles and certificates for app and push notifications"
   lane :certs do
     app_store_connect_api_key(key_filepath: "ios/apollos.p8")
+    match_nuke(type: "development", app_identifier: [ENV["DELIVER_APP_IDENTIFIER"], "#{ENV["DELIVER_APP_IDENTIFIER"]}.OneSignalNotificationServiceExtension"])
+    match_nuke(type: "distribution", app_identifier: [ENV["DELIVER_APP_IDENTIFIER"], "#{ENV["DELIVER_APP_IDENTIFIER"]}.OneSignalNotificationServiceExtension"])
     match(type: 'development', app_identifier: [ENV["DELIVER_APP_IDENTIFIER"], "#{ENV["DELIVER_APP_IDENTIFIER"]}.OneSignalNotificationServiceExtension"],)
     match(type: 'appstore', app_identifier: [ENV["DELIVER_APP_IDENTIFIER"], "#{ENV["DELIVER_APP_IDENTIFIER"]}.OneSignalNotificationServiceExtension"],)
     get_push_certificate(app_identifier: ENV["DELIVER_APP_IDENTIFIER"])
@@ -55,7 +57,20 @@ platform :ios do
     increment_build_number(build_number: number_of_commits, xcodeproj: "ios/" << proj << ".xcodeproj")
     build_app(workspace: "ios/" << proj << ".xcworkspace", scheme: proj)
     changelog_from_git_commits
+    upload_app_privacy_details_to_app_store(
+      app_identifier: ENV["DELIVER_APP_IDENTIFIER"]
+    )
     testflight
+  end
+
+  desc "Promote the most recent TestFlight build to external beta"
+  lane :beta do
+    app_store_connect_api_key(key_filepath: "ios/apollos.p8", in_house: false)
+    testflight(
+      distribute_only: true,
+      distribute_external: true,
+      groups: ["Beta Testers"]
+    )
   end
 
   desc "Promote the most recent Testflight version to production"
@@ -63,10 +78,9 @@ platform :ios do
     app_store_connect_api_key(key_filepath: "ios/apollos.p8", in_house: false)
     deliver(
       skip_binary_upload: true,
-      skip_screenshots: false,
-      skip_metadata: false,
+      overwrite_screenshots: true,
       submit_for_review: true,
-      automatic_release: false,
+      automatic_release: true,
       submission_information: {
         add_id_info_uses_idfa: false
       },
@@ -91,6 +105,38 @@ platform :android do
     upload_to_play_store(
       track: "internal",
       skip_upload_apk: true,
+      json_key: "android/key.json",
+      package_name: package_name
+    )
+  end
+
+  desc "Promote current build to beta"
+  lane :beta do
+    s = File.read("../android/app/build.gradle")
+    package_name = s[/applicationId\s+"([\w+\.]*)"/, 1]
+
+    upload_to_play_store(
+      track: "internal",
+      track_promote_to: "beta",
+      version_code: number_of_commits,
+      skip_upload_apk: true,
+      skip_upload_aab: true,
+      json_key: "android/key.json",
+      package_name: package_name
+    )
+  end
+
+  desc "Promote current build to production"
+  lane :production do
+    s = File.read("../android/app/build.gradle")
+    package_name = s[/applicationId\s+"([\w+\.]*)"/, 1]
+
+    upload_to_play_store(
+      track: "internal",
+      track_promote_to: "production",
+      version_code: number_of_commits,
+      skip_upload_apk: true,
+      skip_upload_aab: true,
       json_key: "android/key.json",
       package_name: package_name
     )
