+++ package.json
@@ -41,22 +41,22 @@
   "dependencies": {
     "@apollo/client": "3.3.20",
     "@apollo/react-hooks": "^3.1.3",
-    "@apollosproject/config": "^2.36.3",
+    "@apollosproject/config": "^2.42.0",
     "@apollosproject/react-native-airplay-btn": "^0.2.0",
-    "@apollosproject/ui-analytics": "^2.36.3",
-    "@apollosproject/ui-auth": "^2.36.3",
-    "@apollosproject/ui-connected": "^2.36.3",
-    "@apollosproject/ui-fragments": "^2.36.3",
-    "@apollosproject/ui-htmlview": "^2.36.3",
-    "@apollosproject/ui-kit": "^2.36.3",
-    "@apollosproject/ui-mapview": "^2.36.3",
-    "@apollosproject/ui-media-player": "^2.36.3",
-    "@apollosproject/ui-notifications": "^2.36.3",
-    "@apollosproject/ui-onboarding": "^2.36.3",
-    "@apollosproject/ui-passes": "^2.36.3",
-    "@apollosproject/ui-prayer": "^2.36.3",
-    "@apollosproject/ui-scripture": "^2.36.3",
-    "@apollosproject/ui-test-utils": "^2.36.3",
+    "@apollosproject/ui-analytics": "^2.42.0",
+    "@apollosproject/ui-auth": "^2.42.0",
+    "@apollosproject/ui-connected": "^2.42.0",
+    "@apollosproject/ui-fragments": "^2.42.0",
+    "@apollosproject/ui-htmlview": "^2.42.0",
+    "@apollosproject/ui-kit": "^2.42.0",
+    "@apollosproject/ui-mapview": "^2.42.0",
+    "@apollosproject/ui-media-player": "^2.42.0",
+    "@apollosproject/ui-notifications": "^2.42.0",
+    "@apollosproject/ui-onboarding": "^2.42.0",
+    "@apollosproject/ui-passes": "^2.42.0",
+    "@apollosproject/ui-prayer": "^2.42.0",
+    "@apollosproject/ui-scripture": "^2.42.0",
+    "@apollosproject/ui-test-utils": "^2.42.0",
     "@bugsnag/react-native": "^7.14.1",
     "@gorhom/bottom-sheet": "3.4.1",
     "@react-native-community/async-storage": "^1.5.1",
@@ -111,10 +111,11 @@
     "recompose": "^0.30.0",
     "rn-fetch-blob": "^0.10.16",
     "stream": "^0.0.2",
-    "yup": "0.26.6"
+    "yup": "0.26.6",
+    "react-native-sha256": "^1.4.7"
   },
   "devDependencies": {
-    "@apollosproject/eslint-config": "^2.36.3",
+    "@apollosproject/eslint-config": "^2.42.0",
     "@babel/core": "^7.14.6",
     "@babel/runtime": "^7.14.6",
     "@bam.tech/react-native-make": "^3.0.3",
