+++ src/client/cache.js
@@ -3,18 +3,6 @@ import AsyncStorage from '@react-native-community/async-storage';
 import { CachePersistor } from 'apollo3-cache-persist';
 import ApollosConfig from '@apollosproject/config';
 
-// NOTE: this file is generated at runtime
-// eslint-disable-next-line
-import introspectionQueryResultData from "./fragmentTypes.json";
-
-const finalPossibleTypes = {};
-introspectionQueryResultData.__schema.types.forEach((supertype) => {
-  if (supertype.possibleTypes) {
-    finalPossibleTypes[supertype.name] = [
-      ...supertype.possibleTypes.map((subtype) => subtype.name),
-    ];
-  }
-});
 // We reset our apollo cache based an env value and static number.
 // In the future, we should also look at resetting the app when an error occurs related to Apollo.
 // You can also increment this number to force a manual reset of the cache.
@@ -25,7 +13,7 @@ const nodeCacheRedirect = (_, { id }, { getCacheKey }) =>
   id ? getCacheKey({ __typename: id.split(':')[0], id }) : null;
 
 const cache = new InMemoryCache({
-  possibleTypes: finalPossibleTypes,
+  possibleTypes: ApollosConfig.TYPEMAP,
   cacheRedirects: {
     Query: {
       node: nodeCacheRedirect,
