+++ src/client/index.js
@@ -1,14 +1,18 @@
 import { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { ApolloProvider, ApolloClient, ApolloLink, gql } from '@apollo/client';
+// import { createPersistedQueryLink } from '@apollo/client/link/persisted-queries';
+// import { sha256 } from 'react-native-sha256';
 import { getVersion, getApplicationName } from 'react-native-device-info';
+import { Platform } from 'react-native';
+import { createUploadLink } from 'apollo-upload-client';
+import ApollosConfig from '@apollosproject/config';
 
 import { authLink, buildErrorLink } from '@apollosproject/ui-auth';
 import { updatePushId } from '@apollosproject/ui-notifications';
 
 import { NavigationService } from '@apollosproject/ui-kit';
 
-import httpLink from './httpLink';
 import cache, { ensureCacheHydration } from './cache';
 
 const wipeData = () =>
@@ -37,9 +41,33 @@ const onAuthError = async () => {
   NavigationService.resetToAuth();
 };
 
-const errorLink = buildErrorLink(onAuthError);
+let uri = ApollosConfig.APP_DATA_URL;
+const androidUri = ApollosConfig.ANDROID_URL || '10.0.2.2';
 
-const link = ApolloLink.from([authLink, errorLink, httpLink]);
+// Android's emulator requires localhost network traffic to go through 10.0.2.2
+if (Platform.OS === 'android') {
+  uri = uri.replace('localhost', androidUri);
+}
+
+const errorLink = buildErrorLink(onAuthError);
+// const apqLink = createPersistedQueryLink({
+//   sha256,
+//   useGETForHashedQueries: true,
+// });
+
+const link = ApolloLink.from([
+  authLink,
+  errorLink,
+  // apqLink,
+  createUploadLink({
+    uri,
+    headers: {
+      ...(ApollosConfig.CHURCH_HEADER
+        ? { 'x-church': ApollosConfig.CHURCH_HEADER }
+        : {}),
+    },
+  }),
+]);
 
 export const client = new ApolloClient({
   link,
