--- src/ui/ActionBar.js
@@ -1,27 +1,0 @@
-import { ActionBar, ActionBarItem } from '@apollosproject/ui-kit';
-import { useNavigation } from '@react-navigation/native';
-import { RockAuthedWebBrowser } from '@apollosproject/ui-connected';
-
-const Toolbar = () => {
-  const navigation = useNavigation();
-  return (
-    <RockAuthedWebBrowser>
-      {(openUrl) => (
-        <ActionBar>
-          <ActionBarItem
-            onPress={() => navigation.navigate('Passes')}
-            icon="check"
-            label="Check-in"
-          />
-          <ActionBarItem
-            onPress={() => openUrl('https://apollosrock.newspring.cc/page/186')}
-            icon="download"
-            label="Give"
-          />
-        </ActionBar>
-      )}
-    </RockAuthedWebBrowser>
-  );
-};
-
-export default Toolbar;
